+++ packages/vendure-plugin-invoices/src/entities/invoice-config.entity.ts
@@ -1,7 +1,6 @@
-import { Column, Entity, ColumnType, getMetadataArgsStorage } from 'typeorm';
 import { DeepPartial, Logger, VendureEntity } from '@vendure/core';
-import { InvoicePlugin } from '../../invoice.plugin';
-import { loggerCtx } from '../../constants';
+import { Column, ColumnType, Entity } from 'typeorm';
+import { loggerCtx } from '../constants';
 
 @Entity('invoice_config')
 export class InvoiceConfigEntity extends VendureEntity {
@@ -14,6 +13,9 @@ export class InvoiceConfigEntity extends VendureEntity {
 
   @Column({ default: false })
   enabled: boolean = false;
+
+  @Column({ default: true })
+  createCreditInvoices: boolean = true;
 
   @Column({ type: resolveTemplateColumnType(), nullable: true })
   templateString?: string | null;