+++ packages/vendure-plugin-invoices/src/scripts/populate-invoice-fields.ts
@@ -1,0 +1,2 @@
+// Find all invoice entities in batches of 100, find corresponding orders and update the invoice fields
+// TODO