+++ packages/vendure-plugin-invoices/src/strategies/google-storage-invoice-strategy.ts
@@ -1,9 +1,6 @@
 import { Logger } from '@vendure/core';
-import { Response } from 'express';
-import { createReadStream, ReadStream } from 'fs';
-import path from 'path';
 import { InvoiceEntity } from '../entities/invoice.entity';
-import { createTempFile, zipFiles, ZippableFile } from '../file.util';
+import { safeRemove } from '../util/file.util';
 import { RemoteStorageStrategy } from './storage-strategy';
 
 interface GoogleInvoiceConfig {
@@ -52,37 +49,18 @@ export class GoogleStorageInvoiceStrategy implements RemoteStorageStrategy {
   async save(
     tmpFile: string,
     invoiceNumber: number,
-    channelToken: string
+    channelToken: string,
+    isCreditInvoice: boolean
   ): Promise<string> {
-    const name = `${channelToken}/${invoiceNumber}.pdf`;
+    let filename = `${invoiceNumber}.pdf`;
+    if (isCreditInvoice) {
+      filename = `${invoiceNumber}-credit.pdf`;
+    }
+    const fullPath = `${channelToken}/${filename}`;
     await this.storage.bucket(this.bucketName).upload(tmpFile, {
-      destination: name,
+      destination: fullPath,
     });
-    return name;
-  }
-
-  async streamMultiple(
-    invoices: InvoiceEntity[],
-    res: Response
-  ): Promise<ReadStream> {
-    res.set({
-      'Content-Type': 'application/zip',
-      'Content-Disposition': `inline; filename="invoices-${invoices.length}.zip"`,
-    });
-    const files: ZippableFile[] = await Promise.all(
-      invoices.map(async (invoice) => {
-        const tmpFile = await createTempFile('.pdf');
-        await this.storage
-          .bucket(this.bucketName)
-          .file(invoice.storageReference)
-          .download({ destination: tmpFile });
-        return {
-          path: tmpFile,
-          name: invoice.invoiceNumber + '.pdf',
-        };
-      })
-    );
-    const zipFile = await zipFiles(files);
-    return createReadStream(zipFile);
+    safeRemove(tmpFile);
+    return fullPath;
   }
 }