+++ packages/vendure-plugin-invoices/src/strategies/s3-storage.strategy.ts
@@ -1,10 +1,7 @@
 import { Logger } from '@vendure/core';
-import { Response } from 'express';
-import { createReadStream, ReadStream } from 'fs';
-import { readFile, writeFile } from 'fs/promises';
-import path from 'path';
+import { readFile } from 'fs/promises';
 import { InvoiceEntity } from '../entities/invoice.entity';
-import { createTempFile, zipFiles, ZippableFile } from '../file.util';
+import { safeRemove } from '../util/file.util';
 import { RemoteStorageStrategy } from './storage-strategy';
 
 export interface Config {
@@ -50,9 +47,14 @@ export class S3StorageStrategy implements RemoteStorageStrategy {
   async save(
     tmpFile: string,
     invoiceNumber: number,
-    channelToken: string
+    channelToken: string,
+    isCreditInvoice: boolean
   ): Promise<string> {
-    const Key: string = `invoices/${channelToken}/${invoiceNumber}.pdf`;
+    let filename = `${invoiceNumber}.pdf`;
+    if (isCreditInvoice) {
+      filename = `${invoiceNumber}-credit.pdf`;
+    }
+    const Key: string = `invoices/${channelToken}/${filename}`;
     await this.s3!.upload({
       Bucket: this.bucket,
       Key,
@@ -60,28 +62,7 @@ export class S3StorageStrategy implements RemoteStorageStrategy {
       ContentType: 'application/pdf',
       ContentDisposition: 'inline',
     }).promise();
+    safeRemove(tmpFile);
     return Key;
-  }
-
-  async streamMultiple(
-    invoices: InvoiceEntity[],
-    res: Response<any, Record<string, any>>
-  ): Promise<ReadStream> {
-    const files: ZippableFile[] = await Promise.all(
-      invoices.map(async (invoice) => {
-        const tmpFile = await createTempFile('.pdf');
-        const object = await this.s3!.getObject({
-          Bucket: this.bucket,
-          Key: invoice.storageReference,
-        }).promise();
-        await writeFile(tmpFile, object.Body?.toString()!);
-        return {
-          path: tmpFile,
-          name: invoice.invoiceNumber + '.pdf',
-        };
-      })
-    );
-    const zipFile = await zipFiles(files);
-    return createReadStream(zipFile);
   }
 }